services/email_deamon/requirements.txt
google_api_python_client==2.160.0
google_auth_oauthlib==1.2.1
protobuf==5.29.3
pymongo==4.6.2
python-dotenv==1.0.1
